feat: app config file
