+++ config/app.config.ts
@@ -1,0 +1,40 @@
+// public routes
+const routes = [
+  {
+    name: 'index',
+    route: '/',
+  },
+  {
+    name: 'program',
+    route: '/program',
+  },
+  {
+    name: 'jobs',
+    route: '/jobs',
+  },
+  {
+    name: 'contact',
+    route: '/contact',
+  },
+]
+
+const adminCategories = [
+  {
+    name: 'companies',
+    icon: 'domain',
+  },
+  {
+    name: 'events',
+    icon: 'calendar',
+  },
+  {
+    name: 'jobs',
+    icon: 'briefcase-outline',
+  },
+  {
+    name: 'users',
+    icon: 'account-outline',
+  },
+]
+
+export { routes, adminCategories }
